chore: add extra file for serializers
