+++ django/restapi/models.py
@@ -1,31 +1,5 @@
-from rest_framework import serializers
-
 from django.contrib.auth.models import User
 from django.db import models
-
-
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ['username', 'password', 'first_name', 'last_name', 'date_joined', 'last_login']
-        # fields = '__all__'
-        extra_kwargs = {
-            'password': {'write_only': True},
-            'first_name': {'required': True},
-            'last_name': {'required': True},
-            'last_login': {'read_only': True},
-            'date_joined': {'read_only': True},
-        }
-
-    def create(self, validated_data):
-        user = User(
-            username=validated_data['username'],
-            first_name=validated_data['first_name'],
-            last_name=validated_data['last_name'],
-        )
-        user.set_password(validated_data['password'])
-        user.save()
-        return user
 
 
 class Company(models.Model):
@@ -39,12 +13,6 @@ class Company(models.Model):
 
     def __str__(self):
         return self.name
-
-
-class CompanySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Company
-        fields = '__all__'
 
 
 class Garden(models.Model):
@@ -73,18 +41,6 @@ class Coordinate(models.Model):
         db_table = 'Coordinates'
 
 
-class CoordinateSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Coordinate
-        exclude = ['garden']
-
-
-class GardenSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Garden
-        fields = '__all__'
-
-
 class Bed(models.Model):
     id = models.IntegerField(primary_key=True)
     uuid = models.CharField(max_length=32, null=True)
@@ -99,12 +55,6 @@ class Bed(models.Model):
         return self.uuid
 
 
-class BedSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Bed
-        fields = '__all__'
-
-
 class CompanyPermission(models.Model):
     id = models.BigAutoField(primary_key=True)
     permission = models.CharField(max_length=1)
@@ -117,12 +67,6 @@ class CompanyPermission(models.Model):
         managed = True
 
 
-class CompanyPermissionSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = CompanyPermission
-        fields = '__all__'
-
-
 class GardenPermission(models.Model):
     id = models.BigAutoField(primary_key=True)
     permission = models.CharField(max_length=1)
@@ -133,9 +77,3 @@ class GardenPermission(models.Model):
         app_label = 'restapi'
         db_table = 'GardenPermission'
         managed = True
-
-
-class GardenPermissionSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = GardenPermission
-        fields = '__all__'
