+++ django/restapi/views/company_gardens.py
@@ -1,9 +1,11 @@
 import base64
+import os
 import re
+from uuid import uuid4
 
 from rest_framework.decorators import api_view
 from restapi.models import *
-from restapi.models import Garden
+from restapi.serializer import CoordinateSerializer
 from restapi.util.auth import isCompanyAdmin, isCompanyUser, isGardenAdmin, isGardenUser
 from restapi.util.company_gardens import *
